app management system
Implemented a basic application management system.

Created an app manifest resource file with the configuration for each app, a built in app manager to populate a list of pre-built apps from a folder.

apps can be registered and associated with extensions while browsing files and opened with the path to the file as an input window data along with the manifest info